Files for Country management

+++ src/main/java/com/powerchord/services/Serviceable.java
@@ -5,7 +5,7 @@ import java.util.List;
 public interface Serviceable<T> {
 	boolean create(T model);
 	List<T> getAll();
-	T getOne(int id);
+	T getOne(Long id);
 	boolean update(T model);
-	boolean delete(int id);
+	boolean delete(Long id);
 }
